[HUDI-684] Abstraction for the log data block format that will be used to save records in the log files.
Log block format is tied to the base file format and no extra configuration is required.

Added HoodieHFileDataBlock which is a HoodieDataBlock saving records in the HFile format. This data block will be used with HFile base file format.
Updated the tests for the Hoodie Log Format to test for all log data formats.